backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.16
langgraph==0.0.20
chromadb==0.4.22
simple-salesforce==1.12.5
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
httpx==0.26.0
websockets==12.0
pytest==7.4.4
pytest-asyncio==0.23.3
